server/server/requirements.txt
alembic==1.4.2
aniso8601==8.0.0
attrs==19.3.0
click==7.1.2
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
flask-restful-swagger==0.20.2
Flask-SQLAlchemy==2.4.3
importlib-metadata==1.6.1
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
psycopg2==2.8.5
py==1.8.1
pyparsing==2.4.7
pytest==5.4.3
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.17
wcwidth==0.2.4
Werkzeug==1.0.1
zipp==3.1.0
